appengine-services/video-intelligence-labelling-service/requirements.txt
Flask==1.0.2
gunicorn==19.9.0
google-cloud-videointelligence==1.8.0
firebase-admin==2.16.0
